Move options reader to the separate namespace
